pin-shortcut: Add the option to AllApps apps in homescreen with taskbar
In the desktop device, where the homescreen shows the taskbar
instead of the hotseat, the context it uses is still launcher.
This means that we have to add the functionality to the launcher
side as well.

For the prediction apps, as originally the only way to pin an app
to the hotseat is drag-and-dropping, the prediction apps is reset
and re-added depending on the dragging and dropping actions.
Therefore we have to add a callback to update the predicted apps
after an app is pinned to the taskbar.

Test: Manual
Flag: com.android.window.flags.enable_pinning_app_with_context_menu
Bug: 375648361
Bug: 418038527
Change-Id: I726cd0a020c642d53b342af32da59f1145d3dae9