Remove taskbar InsetsSizeOverride for IME
There was an original attempt to fully remove this override in [1], to
facilitate hiding the IME navigation bar (implemented in [2]). However,
it lead to a test failure, so in [3] we started using a smaller size for
the override in gesture navigation mode.

With further improvements in [4], the previous test failure should no
longer be an issue, so we can fully remove this override and simplify
the logic, as well as remove some edge cases that were not properly
handled.

This also removes the nav buttons translation for IME, as it's no longer
necessary. This is essentially a revert of [5].

  [1]: I86079cb6670a2ae3b6fa883694f8af81df212408
  [2]: I8793db69fb846046300d5a56b3b0060138ef4cd5
  [3]: I102a8bc1f8869ebbce9f8f1fefa651d49a9538ec
  [4]: I7210fc38b70742ca2e45810641ab44c62dc589c8
  [5]: I5ed25b0ffc08145b2221cc5c960e669cc21b2aa7

Flag: EXEMPT bugfix
Bug: 407507279
Bug: 397328336
Test: atest WindowInsetsControllerTests#testDispatchApplyWindowInsetsCount_ime
  manually verifying IME insets between navigation modes
Change-Id: I5d40b7f99a197ca35d2db196711b0087aa146c74