Open all apps by swiping up on connectedDisplay
On connected displays, it doesn't use the all apps thats in the layout .xml but rather the taskbar all apps. That flow to open is quite different which I needed to inspect how that button is invoked. This button is invoked via TaskbarUIController. This approach requires to use the correct TaskbarActivityContext which we can get by using the displayId from the motionEvent so that all apps is opened only for the connected display. To detect swipe up we can use SingleAxisSwipeDetector.

bug:405437439
Test: manually - https://drive.google.com/file/d/1s5oKTcwdTK5O6tr47Wl37HOT_xo9Kbpc/view?usp=sharing
Flag: com.android.window.flags.enable_taskbar_connected_displays
Change-Id: I61603fd485b24295e873f2f8f1c57a2795b07f4a