Remove a task from recents if it moved to a bubble
Tasks in a bubble should not show up in recents.
If a task moves to a bubble, while in recents, we get a signal that the
activity is restarted. Check if the task has not app bubble flag set. If
it does, remove it from recents list.

Bug: 415942808
Test: manual, have recents open, pick an app that is in recents and has
  an icon in taskbar, launch the app to a bubble from long-press menu
  from taskbar, observe recents entry for the app is removed
Flag: com.android.wm.shell.enable_create_any_bubble
Change-Id: I49ac2afb407691766c8cd80789f7ac3d66e32957