Handle reversing the bubble bar expand animation
When the bubble bar expansion animation is stopped in the middle
and reverses, the bubbles jump out of the bounds of the
bubble bar and animate back to the correct position. This change
ensures that the bubbles continue animating from their position at
the time the animation started reversing.

Bug: 415925442
Flag: com.android.wm.shell.enable_bubble_bar
Test: manual
       - this currently isn't wired so just verify that animations
         still work smoothly in general
Change-Id: Ifcc2ad7fca1584129d07571be0c317ba796a7d23