Fix inconsistent bottom padding in work tab.
Was able to repro when the workUtilityView was collapsed/expanded. When collapsed, the height returned during applyPadding() in mWorkManager.getWorkUtilityView().getHeight() is inconsistent compared to when it is expanded. For consistency, the child view's measuredHeight (workFAB and workScheduler) should be used in calculating the total height.

bug: 410012332
Test: manually.
Flag: EXEMPT bug fix
Change-Id: I624a79b41349e36c83c22c992a1942b649f58c3c