desktop-exploded-view: Added desktop task dismissal re-layout logic
This CL adds a function to remove a task from an existing exploded
layout. When a task is removed, other tasks on the same row are
centered. If the last task on a row is removed, then the whole
row is removed and the other rows are centered.

Flag: com.android.launcher3.enable_desktop_exploded_view
Test: DesktopLayoutUtilsTest
Bug: 353949372

Change-Id: I84f271cae760245d3fca2f3442e31b14d2ebc115