Simplify Taskbar stashing logic for IME
This simplifies the logic used to determine when the Taskbar should be
stashed due to the IME being visible. The new behaviour doesn't check
both pinned and transient (as these are opposites), early returns on
non large-screen devices (as there's no taskbar to stash), and checks
whether the IME is docked regardless of navigation mode or hardware
keyboard presence.

This also fixes the IME docked detection by checking the full window
metrics (instead of the insets the drag layer receives, as this view is
smaller than the IME insets themselves), and takes into account the
navigation bar and display cutout insets as well (as these will affect
the IME insets size).

Flag: EXEMPT bugfix
Test: atest TaskbarStashControllerTest#testAnimatePinnedTaskbar_imeShown_replacesIconsWithHandle
  TaskbarStashControllerTest#testAnimatePinnedTaskbar_imeHidden_replacesHandleWithIcons
  TaskbarStashControllerTest#testAnimatePinnedTaskbar_imeHidden_verifyAnimationDuration
  TaskbarStashControllerTest#testAnimateThreeButtonsTaskbar_imeShown_hidesIconsAndBg
  TaskbarStashControllerTest#testAnimateThreeButtonsTaskbar_imeHidden_showsIconsAndBg
  TaskbarStashControllerTest#testSetSystemGestureInProgress_whileImeShown_unstashesTaskbar
  TaskbarStashControllerTest#testSysuiStateImeShowingInApp_imeNotDocked_notStashedForIme
Bug: 407507279
Change-Id: I75c4991d3bedfa53f1c2a486ae18689d253ab774