Abort Nav Handle Long Press if InputConsumer is switched.
I believe the error flow went like this:

1. Touch navbar.
2. Input consumer switches (not sure exact condition for this).
3. NavHandleLongPressInputConsumer is no longer notified about motion
   (slop) or touch up/cancel, so it triggers after 450ms.

I added a test for this case, as well as a similar one with multiple
touch down events, although that was not the cause of the bug. The new
test failed without this change, and passes with it.

Bug: 424114298
Test: atest NexusLauncherRoboTests:com.android.quickstep.inputconsumers.NavHandleLongPressInputConsumerTest
Flag: EXEMPT bug fix
Change-Id: I635b04452038452103fb966f9397d955462ff7f3