Fix crash for launching app located in invisible desktop from taskbar
Currently, there is launcher crash with below steps:
1. open a fullscreen app, say Gmail
2. create a desk (desk 1)
3. create another desk (desk 2), and open 1 app (say chrome) in desk 2
4. pull up Overview and scroll to the fullscreen app (Gmail) created
   in step 1
5. tap chrome (app of desk 2) from the taskbar

The reason for the crash is:
when we scrolled to the fullscreen app located at the leftmost of the
screen, desk 2 is not visible, thus findTaskViewToLaunch returns null,
as it filters the taskView based on if it is visible. This taskView is
passed to createRecentsWindowAnimator() and is used to check
boolean isQuickSwitch = taskView.isEndQuickSwitchCuj(). since it is
null, it crashes there.

This fix lets the launch of desktop falls back to showDesktopApp if
it is not visible.

Bug: 424190733
Flag: com.android.window.flags.enable_multiple_desktops_frontend
Flag: com.android.window.flags.enable_multiple_desktops_backend
Test: manual
Change-Id: I776b40757a80a07415ff5e6ee7415e320d90db89