pin-shortcut: Let newly pinned app to be at the end of taskbar.
Originally the newly pinned apps are pinned to the first available
space in the hotseat to minimize the change to other items in the
hotseat. However, as the feature is taskbar only, pinning an app
to the middle of the taskbar is not straightforward to users.
This cl changed the behavior so that the hotseat compact to the
left if needed so that all newly pinned apps will be pinned to the
end of the taskbar.

Test: Manual
Flag: com.android.window.flags.enable_pinning_app_with_context_menu
Bug: 375648361
Bug: 424469003
Change-Id: If689d4b2483d0e77802d09c55dacec06a5b72fa0