Migrate 4 more tests to LayoutResource.
This change also includes 2 fixes for issues in tests. Callbacks can be
set per test, and don't have to be NO_OP. Also, SQLiteExceptions are
caught in the LauncherProvider layout reseting code. This is ok since
we are going to show a new DB anyways, and this only happens in tests.

Bug: 379378706
Test: Passed Pre-submit and a post-submit run.
Flag: EXEMPT test changes
Change-Id: I9ecf58a5f32527623eaeec72d6e46e98c6897511