Add sandbox context for Taskbar direct boot mode.
It has its own dagger component to swap out preferences that require
encrypted preferences.

This will not affect any tests. It can only break the build if a change
to Launcher's dagger graph happens to go in right before this one.

We will integrate this with TaskbarManager in a future CL.

Flag: com.android.launcher3.enable_taskbar_for_direct_boot
Bug: 324485921
Test: m NexusLauncher
Change-Id: I4f532cf08b74df6cda024dd32804518b243b3666