Remove taskar stashing systemGestureInProgress
This was originally added in [1] to fix an issue where the taskbar did
not unstash while quick switching between apps, with the IME initially
visible.

This can now be removed, as the issue is no longer reproducible without
this check. The IME will be automatically hidden while quick switching,
and while in recents in general.

  [1]: Ib3823a9bd9aebf923ad481c010c7e8544bcb6a4d

Flag: EXEMPT cleanup
Bug: 407507279
Test: atest TaskbarStashControllerTest
Change-Id: I1e2391243453da25c16a0ddce46c619a2e6554c5