pin-shortcut: Check the package of the task if no component key found in all apps
The bug is that right clicking on some recent tasks won't have
the context menu showing up. This is because the task has a
different component key from the app that is running, and thus the
corresponding app in the all apps is not found.
This cl refers to AppPairsController where they check the package
name in addition to the component key to find the original app of
a task.

Test: Manual
Flag: com.android.window.flags.enable_pinning_app_with_context_menu
Bug: 375648361
Bug: 418039033
Change-Id: I4189623a3ea9135404c04683f1e7a80c57fd8ffb