Remove ScrimView's drawing controller when resetting All Apps.
Previously broken flow:
 - Open All Apps on inner screen
   - Drawing controller is set to the AllApps container
 - Close All Apps (resting state becomes NORMAL)
 - Fold, unlock
   - Because we are still in NORMAL, the drawing controller was not
     updated. Therefore when the ScrimView draws (alongside all other
     views due to Launcher starting), it asks the inner screen's All
     Apps container to render its panel on the scrim.

The fix is to remove the drawing controller when All Apps resets,
thereby avoiding trying to draw on the scrim in the first place. The
drawing controller is updated upon the next transition to All Apps,
as usual.

Bug: 426397329
Test: Manual: https://drive.google.com/file/d/1l9qAIkrgFalcO5DABOEIRX1Po3POYzab/view?usp=sharing&resourcekey=0-ccU6q2fbYE6y87ehIKdbEQ
Flag: com.android.launcher3.all_apps_sheet_for_handheld
Change-Id: I0a4872e6bd668583d08848a8b9c5f0f1389c4f0f