desktop-exploded-view: Fix crop corner radius for desktop tasks
The current corner radius calculation in `TaskViewSimulator` is based
on the transform of the recents view itself and doesn't take into
account that tasks can have individual transforms applied. This in turn
results in a visible jump when tasks are stacked behind the launcher.

Flag: com.android.launcher3.enable_desktop_exploded_view
Test: Manual
Bug: 427292246
Change-Id: I2185e0d38b3c9fcce1189c9f62027ff273215ed5