Fix bubble expanding when rotate was clicked.
The bubble would expand sometimes when the rotate suggestion button was
clicked. This would happen because the bubble events are detected using
a global listener that can not report that the event was handled. The
temporary fix is to see if the rotate suggestion button is visible and
overlapping.

Flag: com.android.wm.shell.bug_rotation_button_cover_bubble
Bug: 424165276
Test: manual - open bubbles and turn off auto rotate
Test: move bubble so it is under the rotate button
Test: rotate the device and click the rotate button
Change-Id: Id2347d70e9bef2440e4c9a5c031a98004354028f