pin-shortcut: Fix the janky pin animation
The janky animation is basically caused from the taskbar window
resize.
A brief context is that the context menu on the taskbar is
implemented on the TaskbarActivityContext instead of a overlay
context. This means that every time a context menu is opened, the
Taskbar window will be resized to fullscreen instead of its
visible size.

If we pin an app using the context menu from the taskbar, what
will happen is that the app transitions and the context menu close
animation will run at the same time, but the menu animation
finishes first and triggers the callback to resize the window
while the apps are still animating.

This cl makes sure that the window is reset to its original size
after both app transitions and the context menu close animation
are done.

Test: Manual
Flag: com.android.window.flags.enable_pinning_app_with_context_menu
Bug: 375648361
Bug: 422756606
Change-Id: I3a75a73e026633c4c95da948e758fae9a6779346