[1/2] WidgetPicker: Show featured only once previews are available.
This is to reduce amount of work that happens in arranging the grid
 content and fade in previews as soon as they are available.

Since placeholders seem to be showing up for very small time, it might
feel like flicker, so not showing it. For expand / collapse we keep the
placeholder but don't show the circular indicator -- instead just keep
it for sizing until preview is loaded while the section expands.

Bug: 408283627
Flag: com.android.launcher3.enable_widget_picker_refactor
Test: Perfetto trace and verify animation manually
Change-Id: I7ed2bd1faff6790fecdddf487a98d0e4471e9b92