Make sure applyDeviceProfile changes the icon size of the taskbar
This was broken in this refactor ag/33824088 . I didn't keept that
part of the code no-op, the idea was to move the code to something
similar to what is expected but the icon size wasn't properly
updated.

I added the TODO(b/430382569), using that pattern would also make it
possible to add test. If we use a factory here we could just make sure
the inputs of the factory produce the same DeviceProfile like we
do with the DumptDeviceProfile tests.

Fix: 426575877
Test: NA
Flag: EXEMPT bug fix
Change-Id: Ifbb50473258fe08f0c82c2dde879de68bb63b70b