Centering preview view using gravity Center
We didn't bother to center the view because it just one view that
matches the parent view, the problem is that the peview view is
bigger than it's parent, this is by design so that the preview
matches the full screen and looks the same as the real Launcher.
We then scale the preview view to match the visualization space.

The problem is that the logic that position a view in FrameLayot
doesn't account for children views bigger than itself so in
RTL the view is not "centered" or it doesn't start where we expect
it to start. By using gravity = CENTER it does account correclty
for bigger views.

Fix: 429081966
Test: GridPreviewTest
Flag: EXEMPT bug fix
Change-Id: I1e75087ca68dd3180b0a9f11d9403eafb892391a