Changing WorkspaceIconProfile to WorkspaceProfile and moving the factories to their own file
Changing the name because the ideas was to have different
WorkspaceProfile classes but the values depend too much on each other
it would be easier if they are in the same class.

There as no change other than the name of the methods when creating
WorkspaceProfileNonResponsiveFactory, no variable was change, just the
location of the methods and the name to remove Icon form the name.

Bug: 419264653
Flag: EXEMPT refactor
Test: DeviceProfileDumpTest
Test: All Image tests
Change-Id: I76628f909f93a79a3f503eeed3ae3e79357c37c2