Made bubble bar stay in the persistent taskbar on the home screen.
When the all-apps grid is swiped up, the bubble bar animates
downward to align with the sliding taskbar. At the end of the drag
gesture (after the finger is released), bubble bar vertical position was
calculated using `translationYForState`, which was not updated.

This change switches to using `bubbleBarTranslationY`, which correctly
represents the bubble bar's actual position with the necessary
adjustments.

Fixes: 429332327
Test: PersistentBubbleStashControllerTest
Flag: com.android.wm.shell.enable_bubble_bar
Change-Id: I16256b3e8ad1408923ed979625b149921a449edf