Taskbar gesture to a task should keep home screen shown on desktop
In the case of desktop mode where home should always be shown, when a
gesture is cancelled, the launcher state should return to
NORMAL because the home screen is shown behind tasks. This change is
specifically for the desktop form factor.

Bug: 428213157
Bug: 423787279
Test: Manual - Click, drag and release on taskbar to ensure home screen
remains shown.
Flag: com.android.window.flags.show_home_behind_desktop

Change-Id: I27e4285b5ca5a8f90f3417fe2c01881741c1c756