Fix TaskbarOverflowTest flakiness
Fixes flaky testKeyboardQuickSwitchLaunchesTaskAsDesktopApp test, which
verifies that activating a task in taskbar overflow UI activates a desk.
When task is activated in the UI, the desk is activated by posting a
task to UI_HELPER_EXECUTOR, which races with the thread where the test
is executed. To fix flakines, flush UI_HELPER_EXECUTOR before letting
the test end (and before mock expectations are vefiried).

Bug: 431964661
Test: atest TaskbarOverflowTest#testKeyboardQuickSwitchLaunchesTaskAsDesktopApp
    --rerun_until_failure 50
Flag: EXEMPT test fix
Change-Id: I128e5d0e9e30a1cbea1bf7a96e7a39ba809c33c9