Don't enter split if second app selection is mid-installation.
Currently, you can use App A to enter split screen selection, and select
App B that is still downloading as the second split app. However, the
split screen request will fail because an app that is still installing
cannot enter split screen.

This CL disallows selecting an app that's mid-installation as the second
app, so that we don't need to start a Transition unnecessarily on the
Shell side. The split instructions view will wiggle to let the user know
to select a different app, as it does when attempting to select an
AppPair as the second split-app.

Bug: 406532956
Test: Manual - verify that app that is still pending installation can't
be selected as the second split-app
Flag: EXEMPT bug fix

Change-Id: Id5a8a3d519de2bb3b59fa3b3d9948403e2368b9a