pin-overflow: Move PinToTaskbarShortcut back to taskbar
Previously we moved the PinToTaskbarShortcut class to popup
package for the Launcher to access this file easier. However,
since the plan is that only QuickstepLauncher will use this class,
and the overflow feature will need more access to the taskbar
environment, this cl moves it back to the taskbar package. This is
especially important to get the list of icons on the "taskbar" in
additional to the icons that can be shown on the "hotseat"

Test: Manual
Flag: com.android.window.flags.enable_overflow_button_for_taskbar_pinned_items
Bug: 375647992
Change-Id: I3d8aeef4505af13130c23143320e01305a9158ca