Fix memory leak in BubbleBarController
BubbleBarController holds a strong references to TaskbarActivityContext.

Since it is also extending the IBubblesListener.Stub, it may not get
cleaned up when the taskbar is recreated.

Create a separate listener class that extends the IBubblesListener.Stub
class. This listener only holds a weak ref to the controller class which
allows the controller to be garbage collected after taskbar is
destroyed.

Bug: 433335007
Flag: EXEMPT, bugfix
Test: existing tests from presubmit
Test: manual, have some bubbles in bubble bar, rotate device, check that
  new bubbles are posted and bubble bar can be used without issues

Change-Id: I6ffb68608a20f2dad01ffe751c23576b2cd064f4