pin-overflow: Remove pin limit and add overflow view.
This cl let users be able to pin more than the number of the
hotseat can accommodate (usually 6). Also refactored the
TaskbarOverflowView to make it work for the pinned items in
additional to the recent tasks, so that the extra pinned items
will be put into the overflow view.
Currently it is a no-op clicking on the pinned items overflow view.

Test: Manual
Flag: com.android.window.flags.enable_overflow_button_for_taskbar_pinned_items
Bug: 375647992
Change-Id: I54790fc9691945e198f50d63131f7fad8d9ccfc4