Fix GestureState mock leak in TaskAnimationManagerTest
RecentsAnimationListener callbacks are posted onto the Main thread asynchronously in RecentsANimationCallbacks.onAnimationFinished. This created a situation where a GestureState mock that was created during TaskAnimationManagerTest#testLauncherDestroyed_whileRecentsAnimationStartPending_finishesAnimation, but wasn't done being used until shortly after the the test case passed and the test test case (TaskViewTest#notShowBorderByDefault) started. For some reason, that allowed the mock to call the real methods, causing the NPE.

To fix this, we will now flush the main thread after running on the main thread in this test.

Flag: EXEMPT bug fix
Fixes: 407542511
Test: TaskViewTest#notShowBorderByDefault
Change-Id: I363540d5d2a0170bc4447564e7acccf2ff7adba5