Alter timing of desktop split select callback.
Rather than perform the desktop split select callback when the
placeholder anim finishes, perform it when a choice resulting in a task
launch is made.

This means we perform the callback when a second app or the placeholder
view's app icon is selected. Cancel option will not call it.

This is to prevent the Cancel option from having unintended consequences
for desktop tasks.

Bug: 419022969
Test: Manual; second app, app icon, and cancel options all result in
expected behavior for desktop apps.
Flag: EXEMPT, minor change
Change-Id: I9fbe08515f6efeb9bd3caa8a71aa23fd606cdfec