[Taskbar] Refactor LauncherState#getVisibleElements() by removing Launcher param.
This is a follow up CL of ag/34958977.

LauncherState#getVisibleElements() is also called outside taskbar. If
taskbar later depends on LauncherInteractor and pass it to this API, it
will force other components to pass LauncherInteractor as well which is
not feasible.

Thus we have to remove Launcher param and just pass LauncherUiState.

Bug: 404636836
Flag: com.android.launcher3.refactor_taskbar_ui_state
Test: presubmit
Change-Id: Ieb4131573f314c50050228def497c082a0edfd74