Fix bubble bar collapsing after creating bubble from All Apps grid
The newly expanded bubble was collapsing immediately after expansion,
making the animation not visually observable. The collapse request
was coming from TaskbarOverlayController in the launcher process.

This change utilizes the version of
TaskbarStashController.updateAndAnimateTransientTaskbar() that
does not stash the bubble bar if a show bubble request was sent to
shell.

To detect the request, the BubbleShowRequestListener interface
was added to BubbleActivityStarter that should be used to start a new
bubble application or a bubble shortcut.

Fixes: 422358463
Test: TaskbarOverlayControllerTest
Flag: EXEMPT bugfix
Change-Id: I5790a8587777a20d00049f61d58b2c845c30b738