Don't show "app info" and "bubble"
When using voice access or switch access it is possible to see these
system shortcuts since they have a different flow compared to the long
click. Instead of referring to the app being a private space icon, we
can utilize iconAnimationDisabled since this is the only case for
privateSpaceIcon.

bug: 432132972
Test: manually video in bug
Flag: EXEMPT bug fix
Change-Id: I0562ead687dca78aa487a4c64f8407153b49cafe