Moving thread number check to the end of the rule.
It was running in the worker thread since our TAPL test's start.

The check was failing because the threads started by an earlier test
that doesn't have this rule (PreemptiveUnfoldTransitionProgressProviderTest)
apparently relies on finalizers to stop the threads, so the finalizers
might not have run yet. But by the end of our Tapl Test, the threads are
gone.

Bug: 437230340
Flag: EXEMPT test fix
Test: presubmit, abtd on <a certain device where it was failing>
Change-Id: I34ce1f34670d87de499b35f7b6632e3e44726cd8