Make hasNavButton in getter for TaskbarFeatureEvaluator
The Problem: Taskbar recreate is not depended on Display Controller now, so by the time display controller gets recreated for navigation mode change. Taskbar is recreated and hasNavButton value gets incorrect value.

The Solution: simply make the hasNavButton into a getter so value is valued on getting and not before.

Test: Presubmit, Manual
Bug: 438749684
Bug: 437956536
Flag: EXEMPT bugfix
Change-Id: I8c44a8e01f09ec7f1abd1e25bfe8bed7419311f9