[Handoff in Taskbar][3/N] Create TaskbarHandoffController
This is part of a series of changes to add Handoff suggestions to the taskbar. This change introduces `TaskbarHandoffController`, which is responsible for holding Handoff suggestions to be displayed in the taskbar. TaskbarHandoffController does this by integrating with Android platform's TaskContinuityManager APIs.

TaskbarHandoffController is composed of two classes
* HandoffSuggestionList manages a raw list of suggestions to display, and can update itself from a list of RemoteTasks
* HandoffSuggestionMetadataLoader asynchonously loads icons for Handoff suggestions.

Once plumbed into TaskbarViewController, TaskbarHandoffController will update the view when the list of tasks changes

Flag: android.companion.enable_task_continuity
Bug: 400970610
Test: Added Unit Testing
Change-Id: I04d41bc9994d664fcd17fc62d427e17695fbc499