Fixing DeviceProfile calculation order for deprecated grids
The deprecated grids scalable and non-scalable calculate
some AllApps values using an non final version of
WorkspaceProfile. With this change we make sure AllApps is
using the final version of WorkspaceProfile.

This change only affects deprecated grids, as you can see
in the test dumps, it only affects non-scalable.

With this change we can remove calculateAndSetWorkspaceVerticalPadding
and recalculateWorkspacePadding.

Bug: 432070502
Bug: 419264653
Flag: EXEMPT refactor
Test: DeviceProfileDumpTest
Test: All Image tests
Change-Id: Id7d85ed48e7ad9d4d848fcf639f47279d490fe11