Fix test isolation issue in testOpenOverviewWithActionPlusTabKeys
If a previous test leaves a desktop task in the recent tasks list, that task will be the first to be focused through the keyboard shortcut. Also, given that a few other tests in this class also need to clear recent tasks, moving this clean up to setUp to keep all test cases isolated.

Flag: EXEMPT test fix
Fixes: 439474390
Test: TaplTestsQuickstep.testOpenOverviewWithActionPlusTabKeys
Change-Id: Ia5bfc6dcc2552ea815a21b07c1de02ee3545d9b3