Add workspace item when URIs are dropped via system-level event.
This CL adds an ITEM_TYPE_FILE_SYSTEM_FILE to the workspace on
system-level drop event. Currently a workspace item is added for only
the first item in the drop payload, and the item is not fully
populated with relevant file-related metadata. Future CLs will continue
to flesh out the user experience.

Demo: http://dr/file/d/1xLPNVudCjEMIPkljZCvcQUECppw81MKK/view
Design: go/system-drag-and-drop-of-files-in-launcher

Bug: 440196348
Flag: com.android.launcher3.enable_system_drag
Test: atest SystemDragControllerImplTest
Test: atest SystemDragIntegrationTest
Test: atest SystemDragListenerTest
Change-Id: Ie0a79a96d066d237b72b6497d31047b8dbbc76fb