Keep fullscreen only if both source and destination are in desktop-first mode
Currently, Taskbar opens an app in fullscreen exceptionally if the
Taskbar’s display is desktop-first mode and the same app instance is in
fullscreen mode.
This CL adds one more condition to that exception; the display where the
same app instance is on is also in desktop-first mode.
It means that if the source display is in touch-first mode while the
destination display is in desktop-first mode (i.e., projected mode
scenario), the app opens in freeform rather than keeping the fullscreen
mode.

Bug: 440200796
Flag: com.android.window.flags.enable_desktop_first_policy_in_lpm
Test: ConnectedDisplayCujSmokeTests#cuj9p
Change-Id: Iad5dff83d2844de727bf02beec46a4366c9a4946