Fix Overview task focus for TOGGLE_WITH_FOCUS
When TOGGLE_WITH_FOCUS command is triggered to open Overview (using Meta+Tab, for example), there was an issue with focusing the current task as the command check was not added to OverviewCommandHelper.updateRecentsViewFocus and the method returned early. This change adds the command to the method to fix the issue.

Bug: 385128447
Bug: 440569868
Test: atest NexusLauncherTests Launcher3Tests
NexusLauncherOutOfProcTests with enable_meta_tab_toogle_in_overview
enabled
Flag: com.android.launcher3.enable_meta_tab_toggle_in_overview

Change-Id: Iaf3d33ab19b21cf24622ef3de1809b5b42487cdb