Animate back to home screen when activating an empty desk in overview
When selecting an empty desk in overview mode, the created animation
gets aborted due to there being no windows to animate in the transition.
Unfortunately this means that the animation is consumed and not run. Fix
this case by going to rest state when the transition is consumed.

Bug: 429173129
Test: Manual - click on emtpy desk in overview, on both primary and external display
Flag: com.android.window.flags.show_home_behind_desktop
Change-Id: I98d3088e22764fa6f18bd02c468947c0b0f16b26