Allow SystemUiProxy to register and save multiple split-select listeners.
With Split-select now possible from both Launcher and Overview, the
launcher process can now have multiple callback listeners for this
animation. To circumvent burdening Shell with this information,
save the listeners on SystemUiProxy and only expose a "tracker"
instance to the Shell side.

This also adds logic on the different SplitFromDesktopController
instances such that it gives up animating if it's not animating
on the right display (e.g. if it does not have mLauncher and try
to animate the main display, give up).

Bug: 438972618
Test: Manual. Have 2nd display, start split-select on main display,
and it no longer starts on the secondary display.
Flag: com.android.window.flags.enable_non_default_display_split_bugfix

Change-Id: I5a74ffb2c4263e05147512e3ef5e45f6875e6092