Add permission check when evaluating feature status
This checks whether the package has `MANAGE_EXTERNAL_STORAGE`
permission required for showing files on the home screen. Note that the
check is lazily performed and cached for reuse. This will require the
user to restart the launcher (likely the device) in order to see home
screen files when granting the permission in Settings. When the user
revokes the permission, the launcher will be restarted automatically.

Bug: 424466810
Flag: com.android.launcher3.show_files_on_home_screen
Test: manual
Change-Id: Ia80686804e20be62d4e33b6eee2f7b71f50db424