Reserve space for search container when creating home screen files.
When creating home screen files, we need to reserve space for the search
container. Failure to do so would result in home screen files occupying
the same cells which would cause them to be discarded during layout.
Note that this is not an issue when injectable model items are enabled.

Bug: 442874353
Flag: com.android.launcher3.show_files_on_home_screen
Test: atest WorkspaceItemProcessorTest#addsRemainingFileSystemItemsThatWereNotPartOfRestore
Change-Id: I590f3ec7b58ad953c81fe74f2754da028c4c7dc6