Add close option to Taskbar popup menu for Desktop Mode
This change adds a "close" option in the long press menu for Taskbar
icons. It is specific to Desktop mode and will close the calling app. In
multi instance situations it closes all open instances.

Bug: 397961861
Test: Launch multiple apps and invoke popup menu in desktop mode. Try
closing apps using the option and observe they close. Ensure there is no
menu option when not in desktop mode.
Flag: EXEMPT BUGFIX

Change-Id: Ide5fdd650a38456a0af25f0353d3be37b15405f6