Resolve issue of widget disappearing
When quickly tapping a widget and going back home, then long pressing
the widget led to the widget disappearing. This is because we cancel
drag, and since we're in preDrag with the recent popup changes we no longer call dispatchDropComplete in DragController#cancelDrag. We could remove the mIsInPreDrag check so we do call dispatchDropComplete, or we could make it so that we don't call cancelDrag if we are in predrag. Here I do the latter

Bug: 442791916
Test: manual
Flag: com.android.launcher3.home_screen_edit_improvements
Change-Id: I7fe1118b538dc817132127595ed7dc1b13dc9dba